First lab
Failed to complete this during the lab class, although I completed it the evening of the lab class, after giving myself a headache for around 4 hours and not sleeping. Basically I had to create an MP3 class with certain attributes such as: Compression rate, duration, name of artist, name of album, name of song and year the song was made. Simple constructor for the class, not much explaining needed. In the lab question I was given 2 tasks: Create a getSize function that returns the size of the MP3 in KB if it is below 1MB, else returns the size in MB (based off of the compression rate and duration), and another one which takes an array of MP3s, and groups the songs from the same albums together. Task 1 was fairly easy and anyone who reads the code for the function should be able to understand it, task 2 however, has given me the biggest headache since I started studying software engineering (other than calculus). I was not allowed to use ArrayList or HashMap for this task. I could not complete task 2 in the lab but after rewriting my code over and over again I finally got it to do what I want it to do. I just got done with it and it is late for me so sorry if I was a bit incoherent with my explanation.